created mc for trailerLocation
